backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
Pillow==11.3.0
pytesseract==0.3.13
openpyxl==3.1.5
pywin32==310; platform_system=="Windows"

SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
passlib==1.7.4
PyJWT==2.9.0
